fix not refreshing bugish behavior
When adding more : in the window button style settings of the applet, an error occurs.
Melange error log: this[buttonName] is not a function
Reproduce: Having the first window maximized and second window unmaximized it does not update the controls, only when maximizing a window it "catches" the signal but cannot get it once it has been maximized already due to the error.